requirements.txt
selenium==3.14.0
beautifulsoup4==4.6.3
python_dateutil==2.7.3
mysql_connector_repackaged==0.3.1
mysql-connector-python==8.0.11
